#681 fix: Quick Play didn't retry on lobby-code collision, crashed unhandled
Quick Play had its own inline P2002 retry check that only looked at
meta.target — Prisma 7 can omit meta.target on P2002, which the
sibling isLobbyCodeConflict() in /api/lobby already worked around.
Quick Play's independent copy never got that fix, so a code collision
with no meta.target rethrew unhandled instead of retrying (confirmed
in both Sentry and Vercel runtime-error tracking, /api/quick-play).

Extracted isLobbyCodeConflict() to lib/lobby.ts and pointed both
/api/lobby and /api/quick-play's retry loops at it.
